requirements.txt
chromadb==0.3.21
langchain==0.0.155
openai==0.27.6
python-dotenv==1.0.0
streamlit==1.22.0
pypdf==3.8.1
tiktoken==0.3.3